refactor: address review findings before friendly-battle merge
- Remove global --test-concurrency=1 from package.json; scope serial
  execution to daemon test files via { concurrency: false }
- Delete docs/pvp/ speculative server specs (unimplemented, out of scope)
- Extract duplicated AsyncQueue into shared src/friendly-battle/async-queue.ts
