--- docs/pvp/game-design/README.md
@@ -1,18 +1,0 @@
-# PvP 게임 설계 문서
-
-상위 문서: [PvP 문서 인덱스](../README.md)
-
-이 섹션은 Tokénmon 온라인 PvP의 **플레이 경험과 규칙**을 정의한다.
-
-## 포함 문서
-
-1. [배틀 포맷](./battle-format.md)
-2. [성장 및 파티 등록](./progression-and-party-registration.md)
-3. [전설 / 환상 / restricted 정책](./special-pokemon-policy.md)
-4. [세대별 ruleset 설계](./generation-rules.md)
-
-## 권장 읽기 순서
-
-- 먼저 [배틀 포맷](./battle-format.md)으로 플레이 감성을 본다.
-- 다음 [성장 및 파티 등록](./progression-and-party-registration.md)으로 로컬 성장과 온라인 사용의 연결 방식을 본다.
-- 그 다음 [전설 / 환상 / restricted 정책](./special-pokemon-policy.md)과 [세대별 ruleset 설계](./generation-rules.md)로 제약과 세대별 차이를 본다.
