--- docs/pvp/game-design/battle-format.md
@@ -1,96 +1,0 @@
-# PvP 배틀 포맷
-
-상위 문서: [PvP 문서 인덱스](../README.md)
-관련 문서: [성장 및 파티 등록](./progression-and-party-registration.md), [특수 포켓몬 정책](./special-pokemon-policy.md), [실시간 배틀 흐름](../server/battle-flow.md)
-
-## 목표
-
-초기 Tokénmon PvP는 래더/대회 포맷보다 **인게임에서 갑자기 트레이너와 조우한 듯한 감성**을 우선한다.
-즉, 유저는 미리 출전 3마리를 고르고 공개 정보를 많이 보는 방식보다, **자신이 키운 풀 파티를 그대로 들고 와서 실시간으로 판단하는 경험**을 하게 된다.
-
-## 핵심 포맷
-
-- 배틀 타입: **싱글 배틀**
-- 파티 크기: **6마리 등록 / 6마리 전원 사용**
-- 선발: **1번 슬롯 자동 선발**
-- 팀 프리뷰: **없음**
-- 상대 백라인 공개: **없음**
-- 교체: **턴 중 자유 교체 가능**
-- 기절 시: **다음 포켓몬을 직접 선택**
-- 타이머: **행동 선택당 45초**
-- 중복 종: **금지**
-
-## 왜 이런 포맷인가
-
-### 1. 인게임 감성을 유지한다
-
-공식 대회식 팀 프리뷰와 선발 선택은 경쟁적으론 좋지만, Tokénmon이 주려는 감성인 “현장에서 즉시 싸움이 걸리는 느낌”과는 거리가 있다.
-초기 친선 PvP에서는 룰의 완전한 공정성보다도 **게임의 정체성 유지**가 더 중요하다.
-
-### 2. 숨은 정보가 긴장감을 만든다
-
-상대 엔트리를 모르는 상태에서 전투를 시작하면, 교체 판단과 기술 선택의 가치가 커진다.
-이 구조는 특히 친선전에서 “예상 못 한 포켓몬이 뒤에서 나온다”는 재미를 만든다.
-
-### 3. 현재 배틀 엔진 구조와도 잘 맞는다
-
-현 구조는 멀티 배틀보다 싱글 전투에 더 자연스럽게 맞는다. 따라서 초기 온라인 PvP는 싱글에 집중하는 것이 구현 난이도와 안정성 측면에서 맞다.
-
-## 레벨 정책
-
-Tokénmon은 로컬 성장의 보람을 살리고 싶기 때문에, 공식 VGC처럼 모든 포켓몬을 일괄 레벨 50으로 만드는 방향은 초기 친선 PvP와 맞지 않는다.
-
-따라서 초기 룰은 다음처럼 간다.
-
-- 화면 표시 레벨: **실제 성장 레벨 그대로 유지**
-- 배틀 계산용 유효 레벨: **50 이후 압축 적용**
-- 유효 레벨 상한: **60**
-
-이 정책은 다음 균형을 노린다.
-
-- 토큰을 많이 써서 열심히 키운 가치가 남는다.
-- 레벨 차이로 상대를 완전히 찍어누르는 문제는 줄인다.
-- 고레벨 성장의 만족감은 남기되, PvP의 재미는 망치지 않는다.
-
-구체적인 압축 공식은 추후 세부 구현에서 정하되, 룰 상 의미는 “50 이후부터는 성장 효율이 완만해지고, 60 이상은 PvP 계산상 동일 상한”이다.
-
-## 행동 규칙
-
-플레이어는 서버에 아래와 같은 종류의 행동만 보낸다.
-
-- 기술 선택
-- 교체 선택
-- 기절 후 다음 포켓몬 선택
-- 항복
-
-클라이언트는 결과를 계산하지 않는다. 실제 우선순위, 대미지, 상태 변화, 승패 판정은 모두 서버가 수행한다.
-
-## 초기 범위에서 제외하는 것
-
-- 래더 점수 시스템
-- 공개 팀 시트
-- 밴/픽 단계
-- 더블 배틀
-- 토너먼트 운영 기능
-- 관전/리플레이 공유의 완성형 버전
-
-이들은 초기 친선 PvP가 안정화된 뒤 별도 규칙 세트로 분리하는 것이 맞다.
-
-## 확정 규칙 요약
-
-| 항목 | 규칙 |
-|---|---|
-| 배틀 형식 | 싱글 |
-| 사용 마릿수 | 6 |
-| 팀 프리뷰 | 없음 |
-| 선발 | 1번 슬롯 |
-| 타이머 | 45초 |
-| 중복 종 | 금지 |
-| 레벨 표시 | 실제 레벨 |
-| 레벨 계산 | 50 이후 압축, 60 상한 |
-
-## 하위 결정이 필요한 문서
-
-- [성장 및 파티 등록](./progression-and-party-registration.md): 어떤 파티가 온라인에서 유효한가
-- [특수 포켓몬 정책](./special-pokemon-policy.md): 전설/환상/restricted 제한
-- [실시간 배틀 흐름](../server/battle-flow.md): 실제 턴 진행과 이벤트 시퀀스
