--- docs/pvp/game-design/progression-and-party-registration.md
@@ -1,118 +1,0 @@
-# 성장 및 파티 등록 구조
-
-상위 문서: [PvP 문서 인덱스](../README.md)
-관련 문서: [배틀 포맷](./battle-format.md), [세대별 ruleset 설계](./generation-rules.md), [치트 대응 정책](../security/anti-cheat.md), [서버 데이터 모델](../server/data-model.md)
-
-## 핵심 원칙
-
-Tokénmon의 온라인 PvP는 **로컬 스토리/성장을 유지**하면서도, 온라인 대전에서는 **서버가 검증한 파티 스냅샷만 사용**하는 구조로 간다.
-
-즉:
-
-- 성장은 로컬에서 한다.
-- 온라인에서 싸울 때는 로컬 데이터를 그대로 실시간 참조하지 않는다.
-- 서버에 등록한 파티 스냅샷을 기준으로 전투한다.
-
-이 구조는 사용자의 감성과 보안을 동시에 잡기 위한 절충안이다.
-
-## 왜 로컬 파티를 그대로 쓰면 안 되는가
-
-현재 로컬 구조는 스토리/로컬 플레이 중심이다.
-
-- `Config.party`는 현재 파티 슬롯 정보다.
-- 파티 크기 기본값은 3이다.
-- 보유 포켓몬 상태는 종 ID 중심으로 관리되는 경향이 강하다.
-
-즉, 현재 로컬 데이터는 “온라인용 확정 출전 파티”를 표현하기에 적합하지 않다.
-따라서 온라인 PvP는 로컬 파티와 별개의 **등록 파티 개념**을 두는 것이 맞다.
-
-## 등록 파티의 정의
-
-등록 파티는 다음 조건을 만족하는 **서버 저장 스냅샷**이다.
-
-- 특정 **세대(gen1~gen9)** 에 속한다.
-- 최대 6마리로 구성된다.
-- 종 중복 금지 규칙을 만족한다.
-- 특수 포켓몬 제한 규칙을 만족한다.
-- 서버 검증 시점에 유효한 성장 상태를 갖는다.
-- 이후 배틀에서는 이 스냅샷을 기준으로 사용된다.
-
-## 세대별 활성 파티
-
-초기 설계는 **세대별로 활성 등록 파티 1개**를 둔다.
-
-예:
-
-- gen1 온라인 파티 1개 활성
-- gen2 온라인 파티 1개 활성
-- gen4 온라인 파티 1개 활성
-
-이 구조를 선택하는 이유는 다음과 같다.
-
-1. 현재 게임이 세대별로 나뉘어 있다.
-2. 세대마다 등장 포켓몬과 밸런스 기준이 다르다.
-3. 서버도 세대별 ruleset을 적용해야 한다.
-4. UX가 단순하다.
-
-추후 필요하면 세대별 다중 슬롯(예: gen4 파티 A/B/C)로 확장할 수 있지만, 초기에는 **세대당 1개 활성**이 가장 단순하다.
-
-## 재등록 정책
-
-재등록은 허용한다. 다만 의미는 “현재 활성 파티를 최신 상태로 덮어쓴다”에 가깝다.
-
-즉 플레이어는:
-
-- 스토리에서 파티를 육성하고
-- 특정 세대 기준으로 온라인 파티를 다시 만들고
-- 서버에 재등록하여 최신 상태로 갱신할 수 있다.
-
-이 정책은 “키운 포켓몬을 가져오고 싶다”는 감성과 잘 맞는다.
-
-## 추천 등록 플로우
-
-1. 플레이어가 특정 세대를 선택한다.
-2. 클라이언트가 로컬 저장에서 해당 세대의 보유 포켓몬을 읽는다.
-3. 플레이어가 온라인용 6마리를 고른다.
-4. 클라이언트가 서버로 등록 요청을 보낸다.
-5. 서버가 ruleset, 중복 종, 특수 포켓몬 제한, 치트 상태를 검증한다.
-6. 검증 통과 시 서버가 해당 세대의 활성 등록 파티 스냅샷을 갱신한다.
-
-## 온라인 파티와 로컬 파티의 관계
-
-둘은 연결되지만 동일한 것은 아니다.
-
-- 로컬 파티: 현재 스토리/로컬 플레이 편성
-- 온라인 파티: 온라인 PvP용 검증 완료 스냅샷
-
-이렇게 분리해야 다음이 가능하다.
-
-- 스토리 진행과 온라인 밸런스를 분리
-- 온라인 시작 시점 검증 단순화
-- 치트/위조 데이터 차단 포인트 확보
-- 향후 시즌제 ruleset 대응
-
-## UX 관점의 규칙
-
-플레이어 입장에서는 “로컬에서 키운 포켓몬을 세대별 온라인 박스에 등록한다”는 개념으로 보이게 하는 것이 좋다.
-
-권장 UX 문구 예시:
-
-- `gen4 온라인 파티 등록`
-- `현재 gen4 친선전 파티 갱신`
-- `온라인 배틀용으로 이 파티를 등록`
-
-## 설계 결론
-
-초기 온라인 PvP는 다음 구조로 고정한다.
-
-- 로컬 성장 유지
-- 온라인은 서버 등록 스냅샷 사용
-- 세대별 활성 등록 파티 1개
-- 재등록 허용
-- 로컬 `config.party`는 온라인 파티로 재사용하지 않음
-
-## 다음에 읽을 문서
-
-- [세대별 ruleset 설계](./generation-rules.md)
-- [서버 데이터 모델](../server/data-model.md)
-- [치트 대응 정책](../security/anti-cheat.md)
