--- docs/pvp/game-design/special-pokemon-policy.md
@@ -1,107 +1,0 @@
-# 전설 / 환상 / restricted 정책
-
-상위 문서: [PvP 문서 인덱스](../README.md)
-관련 문서: [배틀 포맷](./battle-format.md), [세대별 ruleset 설계](./generation-rules.md), [치트 대응 정책](../security/anti-cheat.md)
-
-## 목표
-
-Tokénmon은 전설과 환상을 아예 금지하는 방향이 아니라, **쓸 수는 있지만 판을 망치지 않도록 제한하는 방향**으로 간다.
-
-이는 다음 두 가지를 동시에 만족시키기 위한 결정이다.
-
-1. 유저가 공들여 얻은 특수 포켓몬의 가치를 살린다.
-2. 친선전이 특정 최상위 포켓몬 몇 마리만의 싸움이 되지 않도록 막는다.
-
-## 기본 분류
-
-초기 온라인 룰에서 포켓몬 관련 분류는 세 층으로 나뉜다.
-
-### 1. 세계관/수집 분류
-
-- `legendary`
-- `mythical`
-
-이 분류는 로컬 게임/도감/수집 의미를 유지한다.
-
-### 2. 온라인 밸런스 분류
-
-- `restricted`
-
-이 분류는 온라인 PvP 밸런스를 위한 별도 개념이다.
-즉, 어떤 포켓몬이 `legendary`이면서 `restricted`일 수도 있고, `mythical`이지만 `restricted`가 아닐 수도 있다.
-
-## 제한 규칙
-
-초기 친선 PvP의 기본 규칙은 다음과 같다.
-
-- **전설 + 환상 총합 최대 2마리**
-- 그중 **restricted 최대 1마리**
-
-예시:
-
-- 가능: 일반 4 + legendary 1 + mythical 1
-- 가능: 일반 5 + restricted 1
-- 불가: restricted 2
-- 불가: legendary/mythical 총합 3
-
-## 왜 환상을 허용하는가
-
-공식 포켓몬 대회에서는 환상이 종종 금지되지만, Tokénmon은 전제가 다르다.
-
-- Tokénmon에서는 환상을 “언제나 획득 가능한 특수 포켓몬”으로 운영할 수 있다.
-- 따라서 접근성 문제 때문에 환상을 일괄 금지할 필요가 없다.
-- 대신 밸런스 측면만 별도로 조정하면 된다.
-
-즉, Tokénmon에서 환상은 “이벤트 한정 희귀 개체”라기보다 “특수하지만 사용 가능한 카테고리”에 더 가깝다.
-
-## 왜 restricted를 따로 두는가
-
-`legendary`와 `mythical`만으로는 실제 PvP 영향력을 충분히 설명하지 못한다.
-
-예를 들어:
-
-- 어떤 legendary는 분위기용/유틸 성격일 수 있다.
-- 어떤 mythical은 충분히 강하지만 판을 혼자 지배하지는 않을 수 있다.
-- 반대로 어떤 포켓몬은 능력/스탯/기술 조합 때문에 한 마리만으로도 메타를 심하게 왜곡할 수 있다.
-
-그래서 온라인 PvP는 “세계관 태그”와 “제한 태그”를 분리해야 한다.
-
-## restricted 판정 기준
-
-특정 포켓몬이 restricted가 되는 기준은 다음과 같다.
-
-1. 단독 캐리력이 지나치게 높다.
-2. 일반 파티 대비 대처 비용이 과도하다.
-3. 숨은 정보 기반 친선전 포맷에서 압박이 지나치다.
-4. 현재 Tokénmon 구현(예: 특성 부재) 때문에 원작보다 더 강해진다.
-
-특히 Tokénmon은 현재 원작과 완전히 동일한 시스템이 아니므로, restricted 선정도 **원작 대회 리스트를 복사하지 않고 TKM 기준으로 재평가**해야 한다.
-
-## Regigigas 같은 예외
-
-원작에서는 특정 특성 때문에 약점이 있는 포켓몬도, Tokénmon 구현 상태에 따라 그 약점이 사라질 수 있다.
-이 경우 원작 평가와 달리 Tokénmon에서는 restricted로 올려야 할 수 있다.
-
-즉 restricted는 “포켓몬 본가 메타의 명성”이 아니라, **현재 Tokénmon 엔진 기준 실전 영향력**으로 결정한다.
-
-## 운영 원칙
-
-- `legendary`, `mythical` 태그는 기존 데이터 의미를 유지한다.
-- `restricted`는 온라인 PvP 전용 정책 데이터로 관리한다.
-- restricted 목록은 **세대별로 따로 관리**한다.
-- 초기에는 시드 리스트로 시작하고, 실제 플레이 데이터를 보며 조정한다.
-
-## 현재 정책 요약
-
-| 규칙 | 값 |
-|---|---|
-| legendary + mythical 총합 | 최대 2 |
-| restricted | 최대 1 |
-| mythical | 허용 |
-| restricted 관리 단위 | 세대별 |
-| restricted 의미 | 온라인 PvP 밸런스 태그 |
-
-## 다음 문서
-
-- [세대별 ruleset 설계](./generation-rules.md)
-- [치트 대응 정책](../security/anti-cheat.md)
