--- docs/pvp/implementation/prd.md
@@ -1,164 +1,0 @@
-# PvP 초기 구현 PRD
-
-상위 문서: [PvP 구현 계획 문서](./README.md)
-기반 설계 문서: [PvP 문서 인덱스](../README.md), [배틀 포맷](../game-design/battle-format.md), [서버 아키텍처](../server/architecture.md), [치트 대응 정책](../security/anti-cheat.md)
-
-## 1. 문제 정의
-
-Tokénmon은 현재 로컬/스토리 중심 구조이기 때문에, 플레이어가 서로 직접 붙는 온라인 친선 PvP 경험이 없다.
-하지만 사용자는 다음을 동시에 원한다.
-
-1. 로컬에서 키운 파티를 온라인에 가져오고 싶다.
-2. 인게임에서 NPC 트레이너를 조우한 듯한 감성으로 싸우고 싶다.
-3. 전투 결과 위조와 성장 상태 위조는 최대한 막고 싶다.
-
-따라서 초기 PvP는 **서버 권한 전투 + 세대별 등록 파티 + 숨은 정보 기반 친선전**을 최소 제품으로 삼는다.
-
-## 2. 제품 목표
-
-### 핵심 목표
-
-- 플레이어가 세대별 온라인 파티를 등록할 수 있다.
-- 친구 코드/룸 코드 기반으로 친선 PvP 룸을 열고 참가할 수 있다.
-- 배틀은 서버가 계산하고, 클라이언트는 명령만 보낸다.
-- 상대 엔트리를 모르는 6v6 싱글 배틀을 실시간으로 진행할 수 있다.
-- 최소한 전투 결과 위조와 치트 오염 저장의 온라인 사용을 막는다.
-
-### 성공 기준
-
-초기 릴리즈에서 다음이 되면 성공으로 본다.
-
-- 플레이어 A/B가 같은 세대 ruleset으로 룸에 들어간다.
-- 서버가 양측 활성 등록 파티를 검증하고 배틀을 시작한다.
-- 각 턴에서 move / switch / replacement / forfeit 명령이 정상 처리된다.
-- 배틀 종료 후 서버에 승패와 이벤트 로그가 남는다.
-- 치트 오염 상태 저장은 온라인 진입이 거부된다.
-
-## 3. 비목표
-
-초기 버전에서 아래는 하지 않는다.
-
-- 래더 / MMR / 매치메이킹
-- 관전 / 리플레이 완성형 기능
-- 공개 팀 시트
-- 밴픽 / 토너먼트 브래킷
-- 더블 배틀
-- 서버 주도 성장 시스템 전체 이관
-
-## 4. 대상 사용자
-
-### 1차 사용자
-- Tokénmon을 이미 플레이하고 있는 기존 유저
-- 로컬에서 키운 파티를 친구와 붙여 보고 싶은 유저
-- 경쟁 e스포츠보다 인게임 감성을 더 중요하게 보는 유저
-
-### 2차 사용자
-- 이후 시즌/래더/관전 기능이 들어오면 경쟁 플레이를 원하는 유저
-
-## 5. 사용자 경험 원칙
-
-- UI/흐름은 “대회 참가”보다 “게임 안에서 바로 트레이너 배틀 시작”에 가깝다.
-- 파티 프리뷰 없이 바로 시작한다.
-- 1번 슬롯이 선발이다.
-- 기절하면 다음 포켓몬을 직접 고른다.
-- 상대의 백라인은 끝까지 비공개다.
-- 실제 레벨은 보이되, PvP 계산용 레벨은 압축한다.
-
-## 6. 기능 요구사항
-
-### F1. 세대별 ruleset 조회
-- 클라이언트가 현재 세대의 온라인 룰을 조회할 수 있어야 한다.
-- ruleset에는 party size, duplicate clause, legendary/mythical limit, restricted limit, effective level cap 등이 포함된다.
-
-### F2. 세대별 온라인 파티 등록
-- 플레이어는 특정 세대에 대해 온라인 파티를 등록/갱신할 수 있어야 한다.
-- 서버는 파티의 유효성을 검증해야 한다.
-- 초기 구조는 세대당 활성 파티 1개다.
-
-### F3. 룸 생성/참가
-- 플레이어 A가 룸을 생성한다.
-- 플레이어 B가 룸 코드로 참가한다.
-- 룸 생성 시 generation / ruleset이 확정된다.
-
-### F4. 서버 권한 배틀 진행
-- 클라이언트는 명령만 제출한다.
-- 서버가 전 턴 결과를 계산한다.
-- 서버는 플레이어별 가시 정보만 내려준다.
-
-### F5. 치트 오염 상태 차단
-- 치트 사용 이력이 있는 저장은 온라인 등록 또는 온라인 진입 시 거부된다.
-- 최소 기준은 현재 코드베이스의 치트 플래그/로그를 반영하는 것이다.
-
-### F6. 배틀 로그 저장
-- 서버는 명령 기록과 결과 이벤트 로그를 저장해야 한다.
-- 최소한 재접속 복구와 사후 디버깅이 가능해야 한다.
-
-## 7. 규칙 요구사항
-
-- 싱글 배틀
-- 6마리 등록 / 6마리 전원 사용
-- 팀 프리뷰 없음
-- 1번 슬롯 자동 선발
-- 중복 종 금지
-- legendary + mythical 총 2 제한
-- restricted 최대 1 제한
-- 실제 레벨 표시 유지
-- 50 이후 계산용 레벨 압축
-- 유효 레벨 최대 60
-- 선택 시간 기본 45초
-
-## 8. 기술 요구사항
-
-### 서버 측
-- party registration API
-- room service
-- realtime gateway (WebSocket)
-- server-authoritative battle orchestration
-- persistence for room/commands/events
-
-### 클라이언트 측
-- ruleset 조회
-- 온라인 파티 등록 UX
-- 룸 생성/참가 UX
-- 실시간 배틀 입력 UX
-- 이벤트 렌더링 UX
-
-## 9. 리스크
-
-### 리스크 A. 로컬 성장 데이터 신뢰성 부족
-로컬 기반 성장 구조 때문에 완전한 무결성 보장은 어렵다.
-
-대응:
-- 서버 등록 스냅샷 사용
-- 치트 오염 저장 차단
-- 배틀 중 실시간 로컬 상태 참조 금지
-
-### 리스크 B. 숨은 정보 처리 실수
-서버가 잘못 구현되면 상대 백라인이 누출될 수 있다.
-
-대응:
-- 플레이어별 view projection 구조 강제
-- event/schema 레벨에서 public/private payload 분리
-
-### 리스크 C. 재접속/타임아웃 처리 복잡도
-실시간이므로 네트워크 이슈가 필연적이다.
-
-대응:
-- 초기에는 친구전 중심으로 제한
-- room snapshot + battle event log 기반 복구
-- 타임아웃 정책은 단순하게 시작
-
-## 10. 수용 기준
-
-다음 시나리오가 전부 통과하면 MVP 수용 가능으로 본다.
-
-1. gen4 파티 등록 성공
-2. 제한 위반 파티 등록 실패
-3. 치트 오염 저장 온라인 등록 실패
-4. 룸 생성 및 코드 참가 성공
-5. 배틀 시작 후 양측 1번 슬롯 선발 확인
-6. move/move 턴 처리 성공
-7. switch/move 턴 처리 성공
-8. faint 후 replacement 요청 성공
-9. 배틀 종료 및 승패 기록 저장 성공
-10. 재접속 시 현재 공개 상태 복구 성공
