--- docs/pvp/roadmap/README.md
@@ -1,15 +1,0 @@
-# PvP 로드맵 문서
-
-상위 문서: [PvP 문서 인덱스](../README.md)
-
-이 섹션은 Tokénmon 온라인 PvP를 **어떤 순서로 구현할지** 정리한다.
-
-## 포함 문서
-
-1. [구현 로드맵](./rollout-plan.md)
-
-## 목적
-
-- 설계를 바로 구현 가능한 단계로 분해
-- 초기 친선 PvP에서 꼭 필요한 범위 우선 정리
-- 이후 래더/관전/리플레이 확장 여지 확보
