--- docs/pvp/security/README.md
@@ -1,15 +1,0 @@
-# PvP 보안 / 치트 대응 문서
-
-상위 문서: [PvP 문서 인덱스](../README.md)
-
-이 섹션은 Tokénmon 온라인 PvP의 **신뢰성, 부정행위 방지, 운영 기준**을 정의한다.
-
-## 포함 문서
-
-1. [치트 대응 정책](./anti-cheat.md)
-
-## 핵심 초점
-
-- 전투 결과 위조 방지
-- 파티 성장 상태 위조 억제
-- 친선전 수준에서 현실적인 보안 절충 정의
