--- docs/pvp/security/anti-cheat.md
@@ -1,103 +1,0 @@
-# 치트 대응 정책
-
-상위 문서: [PvP 문서 인덱스](../README.md)
-관련 문서: [서버 아키텍처](../server/architecture.md), [성장 및 파티 등록](../game-design/progression-and-party-registration.md), [API 계약](../server/api-contract.md), [구현 로드맵](../roadmap/rollout-plan.md)
-
-## 가장 막고 싶은 것
-
-초기 PvP에서 최우선으로 막아야 할 것은 다음 두 가지다.
-
-1. **전투 결과 위조**
-2. **자신의 파티 성장 상태 위조**
-
-이 두 가지를 막지 못하면 온라인 PvP는 금방 신뢰를 잃는다.
-
-## 핵심 방어 원칙
-
-### 1. 전투는 서버 권한
-
-가장 중요한 원칙이다.
-
-- 클라이언트는 결과를 계산하지 않는다.
-- 서버만 승패를 판정한다.
-- 서버만 배틀 로그를 확정한다.
-
-이렇게 해야 클라이언트가 “내가 이겼다”고 주장하는 구조를 원천적으로 제거할 수 있다.
-
-### 2. 온라인은 등록 스냅샷 기반
-
-배틀 직전 로컬 메모리를 그대로 읽어 싸우게 하면 조작 여지가 너무 크다.
-따라서 서버가 검증해 저장한 **온라인 파티 스냅샷**을 기준으로 싸워야 한다.
-
-### 3. 치트 오염 상태는 온라인 불가
-
-현재 코드베이스에는 치트 관련 진입점이 존재한다. 초기 PvP에서는 최소한 다음 중 하나가 필요하다.
-
-- 치트 사용 이력이 있는 저장은 온라인 불가
-- 또는 온라인 전용 프로필/저장을 별도로 강제
-
-초기 버전 기준으로는 **치트 오염 저장 온라인 불가**가 가장 단순하고 명확하다.
-
-## 성장 상태 위조를 막는 방법
-
-완전한 방지는 어렵지만, 아래 조합으로 상당 부분 줄일 수 있다.
-
-- 등록 시 서버 검증
-- 등록 스냅샷 사용
-- 등록 이후 배틀 중 실시간 로컬 상태 참조 금지
-- ruleset 기준 필터링
-- 온라인 허용 저장 상태 검사
-
-향후에는 추가로 다음을 고려할 수 있다.
-
-- 서버 서명 기반 registration proof
-- 로컬 저장 해시/검증
-- 이벤트 기반 성장 로그 업로드
-
-## 현실적인 절충
-
-스토리/성장이 로컬에 있는 구조에서는 완전 무결성을 바로 달성하기 어렵다.
-따라서 초기 친선 PvP의 현실적인 목표는 다음이다.
-
-- 결과 위조는 강하게 차단
-- 성장 상태 위조는 등록 단계에서 크게 억제
-- 남는 위험은 친선전 수준에서 수용 가능 범위로 축소
-
-즉, 초기 PvP는 “완전 무결한 e스포츠 환경”이 아니라, **친선전으로 충분히 신뢰 가능한 수준**을 목표로 한다.
-
-## 온라인 진입 조건 제안
-
-플레이어는 다음 조건을 만족해야 온라인 PvP에 진입할 수 있다.
-
-- 해당 세대의 활성 등록 파티가 존재함
-- 등록 파티가 현재 ruleset 검증을 통과함
-- 치트 오염 상태가 아님
-- 서버 인증 세션이 유효함
-
-## 구현 우선순위
-
-### 반드시 초기부터 넣어야 하는 것
-- 서버 권한 배틀
-- 등록 파티 스냅샷
-- 룰 검증
-- 치트 오염 저장 차단
-- 배틀 이벤트 로그 저장
-
-### 추후 강화 가능한 것
-- 재접속 악용 방지
-- 부정행위 탐지 로그 분석
-- 관전/리플레이 무결성 보장
-- 서버 기반 성장 검증 강화
-
-## 설계 결론
-
-초기 PvP의 보안은 “무조건 모든 치트를 완벽히 막는다”가 아니라, **가장 치명적인 두 위조를 먼저 막는 구조를 고정한다**가 핵심이다.
-그 구조의 중심은 다음 두 가지다.
-
-- 서버 권한 전투
-- 서버 검증 온라인 파티 등록
-
-## 다음 문서
-
-- [구현 로드맵](../roadmap/rollout-plan.md)
-- [API 계약](../server/api-contract.md)
