--- docs/pvp/server/architecture.md
@@ -1,120 +1,0 @@
-# PvP 서버 아키텍처
-
-상위 문서: [PvP 문서 인덱스](../README.md)
-관련 문서: [서버 데이터 모델](./data-model.md), [API 계약](./api-contract.md), [실시간 배틀 흐름](./battle-flow.md), [치트 대응 정책](../security/anti-cheat.md)
-
-## 핵심 결정
-
-초기 온라인 PvP는 **서버 권한(server-authoritative)** 구조로 간다.
-
-즉:
-
-- 클라이언트는 명령만 보낸다.
-- 서버가 배틀 로직 전체를 계산한다.
-- 서버가 결과를 저장한다.
-- 클라이언트는 서버 이벤트를 렌더링한다.
-
-이 결정은 Tokénmon PvP에서 가장 중요한 보안 요구사항 두 개를 직접 겨냥한다.
-
-1. 전투 결과 위조 방지
-2. 성장 상태 위조 최소화
-
-## 왜 서버 권한이어야 하는가
-
-만약 클라이언트가 결과를 계산하면 다음 문제가 생긴다.
-
-- 대미지 계산 위조
-- 승패 위조
-- 상태 이상/교체 흐름 위조
-- 경험치/보상 결과 위조
-- 로컬 저장 조작을 통한 파티 상태 위조
-
-반면 서버 권한 구조에서는 클라이언트가 “이 턴에 무엇을 하겠다”만 제안하고, 실제 결과는 서버가 결정하므로 위조 가능성이 훨씬 줄어든다.
-
-## 권장 구성 요소
-
-### 1. Party Registration Service
-
-- 세대별 온라인 파티 등록
-- 파티 유효성 검증
-- 치트 상태 검증
-- 활성 파티 갱신
-
-### 2. Battle Room Service
-
-- 룸 생성/참가
-- 초대 코드 관리
-- 플레이어 입장 상태 관리
-- 세대 및 ruleset 확정
-
-### 3. Battle Simulation Engine
-
-- 기술 선택 처리
-- 교체 처리
-- 속도/우선도/판정
-- 대미지 계산
-- 상태 변화
-- 승패 판정
-
-### 4. Event Stream / Realtime Gateway
-
-- WebSocket 연결 유지
-- 턴 요청 이벤트 송신
-- 턴 결과 이벤트 송신
-- 재접속 처리
-
-### 5. Persistence Layer
-
-- 등록 파티 스냅샷 저장
-- 룸 상태 저장
-- 배틀 커맨드 저장
-- 배틀 이벤트 로그 저장
-
-## 통신 모델
-
-권장 구조는 **HTTP + WebSocket 혼합**이다.
-
-- HTTP: 등록, 룸 생성, 룸 참가, 메타 조회
-- WebSocket: 실시간 턴 진행, 상태 동기화, 배틀 이벤트 수신
-
-이 조합이 좋은 이유:
-
-- 초기 구현이 단순하다.
-- 실시간성과 요청/응답형 API를 분리할 수 있다.
-- 재접속/이벤트 스트림 관리가 용이하다.
-
-## 숨은 정보 처리
-
-초기 PvP는 팀 프리뷰가 없고 상대 백라인도 비공개다. 따라서 서버는 플레이어별로 다른 정보를 내려줘야 한다.
-
-예:
-
-- 내 파티 전체 상태는 전부 보인다.
-- 상대는 현재 필드에 나온 포켓몬 정보 중심으로만 보인다.
-- 상대 백라인은 서버 내부에만 존재한다.
-
-즉, 서버는 단순 브로드캐스트가 아니라 **플레이어별 시야(view projection)** 를 만들어야 한다.
-
-## 재접속 고려
-
-실시간 PvP에서는 네트워크 이슈가 반드시 생긴다. 초기 버전에서도 최소한 다음은 고려해야 한다.
-
-- 룸 상태 복구
-- 마지막 공개 정보 재전송
-- 남은 선택 시간 재계산
-- 이미 제출한 명령 여부 확인
-
-## 아키텍처 결론
-
-초기 PvP 서버는 다음 원칙으로 구현한다.
-
-- 서버가 유일한 정답 소스(source of truth)
-- 클라이언트는 입력 장치 + 렌더러
-- 숨은 정보는 서버에서 플레이어별로 분리 투영
-- 룸/파티/배틀 로그를 데이터 모델로 명확히 분리
-
-## 다음 문서
-
-- [서버 데이터 모델](./data-model.md)
-- [API 계약](./api-contract.md)
-- [실시간 배틀 흐름](./battle-flow.md)
